pushing coments and clean up to vectors
