+++ helixdb/src/helix_engine/vector_core/mod.rs
@@ -1,6 +1,7 @@
 pub mod vector;
 pub mod hnsw;
 pub mod vector_core;
+pub mod vector_distance;
 
 #[cfg(test)]
 mod vector_tests;
